LayoutTests/fast/animation/request-animation-frame-missing-arguments.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test how animation methods react to too few arguments.');
shouldThrow("webkitRequestAnimationFrame()");
shouldThrow("webkitCancelAnimationFrame()");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
